Refactor tree manager into exec and edit components
